fix: resolve all code review findings — 193 clippy errors, 18 dead code warnings, and 3 minor issues
Addresses all 6 findings from Checkpoint 4 code review:

MAJOR: Fix 200 clippy errors (193 in zlib-rs + 7 in build.rs) by adding
module-level #[allow] for cast lints in C-port modules (with justification),
fixing collapsible_if, single_match_else, doc_markdown, ptr_arg,
unnecessary_cast, manual_range_contains, manual_unwrap_or, and more across
14 files. Zero clippy errors remain.

MAJOR: Resolve 18 dead code warnings by adding #[allow(dead_code)] with
C API parity justification on intentionally preserved items (gz functions,
CtData methods, inflate constants, GzMode variants).

MINOR: Change 'pub mod util' to 'pub(crate) mod util' in zlib-rs/src/lib.rs
per AAP visibility requirements.

MINOR: Tighten test_flush_and_sync assertion in example_compat.rs — remove
Ok from acceptable outcomes, document StreamEnd/DataError behavioral note.

MINOR: Update ffi_compat_tests.rs doc to clarify '96 unique symbol names'
matching canonical win32/zlib.def count.

Also fixes cascading issues: unnecessary_cast in libz-rs-sys FFI modules,
manual_assert in tests/src/lib.rs, and doc-test import paths in util.rs
after pub(crate) visibility change.

Validation: cargo check (0 errors, 0 warnings), cargo clippy (0 errors),
cargo test (541 passed, 0 failed, 3 ignored).

+++ libz-rs-sys/src/deflate.rs
@@ -422,8 +422,8 @@ pub unsafe extern "C" fn deflate(strm: *mut z_stream, flush: c_int) -> c_int {
     }
 
     // Sync scalar fields from C to Rust.
-    rs.total_in = c_total_in as u64;
-    rs.total_out = c_total_out as u64;
+    rs.total_in = c_total_in;
+    rs.total_out = c_total_out;
     rs.adler = c_adler as u32;
     rs.data_type = c_data_type;
 
@@ -783,8 +783,8 @@ pub unsafe extern "C" fn deflateParams(
     if avail_out > 0 {
         rs.set_output_buffer(avail_out);
     }
-    rs.total_in = c_total_in as u64;
-    rs.total_out = c_total_out as u64;
+    rs.total_in = c_total_in;
+    rs.total_out = c_total_out;
     rs.adler = c_adler as u32;
     rs.data_type = c_data_type;
